# Cutadapt: Speed up module by caching parsing versions