<!--
Many thanks to contributing to MultiQC!
Please fill in the appropriate checklist below (deleete whatever is not relevant).
-->

- [x] This comment contains a description of changes (with reason)

The cutadapt module parses versions. Ironically, this version parsing is most of the module execution time. I know that modules shouldn't focus too heavily on performance, but for such low hanging fruit and such disproportionate performance costs that does not count. Plus I love cutadapt. It is the best adapter trimmer hands down. It can be used for ONT as well. As a result almost every MultiQC invocation should find some cutadapt reports. On the production test data you provided to me, this small change makes quite a difference.

Before:
```
|           multiqc |  - 10.33s: Searching files
|           multiqc |  - 9.42s: Running modules
|           multiqc |  - 4.94s: Compressing report data
```

After:
```
|           multiqc |  - 10.20s: Searching files
|           multiqc |  - 6.73s: Running modules
|           multiqc |  - 4.94s: Compressing report data
```